# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/audit-logs": {
      "get": {
        "operationId": "AuditController_list",
        "summary": "查询审计日志（支持按人/资源/动作/时间过滤）",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "审计日志"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/audit-logs/{resource}/{id}/timeline": {
      "get": {
        "operationId": "AuditController_timeline",
        "summary": "某条业务资源的完整操作记录（详情页操作记录）",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "审计日志"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "CreatorOps · 达人合作管理平台 API",
    "description": "聚猩智媒内部达人合作全流程管理平台后端接口。\n\n**鉴权**：除 /api/auth/login、/api/auth/refresh、/api/health 外，所有接口都需要 Bearer Token。\n\n**统一响应结构**：成功 `{ success: true, data, requestId, timestamp }`；\n失败 `{ success: false, code, message, details?, path, requestId, timestamp }`。\n\n**金额与比率约定**：金额返回为字符串（元，2 位小数）以避免浮点误差；\n分润比率以基点(bp)整数传输，1000 = 10%。\n\n**权限模型**：RBAC + 数据范围（ALL/TEAM/OWN）+ 资源归属校验，见 docs/permissions.md。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "当前环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "登录后获取的 accessToken"
      }
    },
...
  }
}